+++ CHANGELOG.md
@@ -22,7 +22,9 @@
   - `neuron rib` is renamed to `neuron gen` (neuron has migrated from rib/shake to reflex)
   - `neuron search`: full-text search uses exact matching (#526)
   - `neuron search`: highlight and focus matching line in `bat` preview (#556)
-  - `neuron query`: query CLI interface has been changed (see docs)
+  - `neuron query`: 
+    - query CLI interface has been changed (see docs)
+    - JSON output is now pretty printed (using `aeson-pretty`)
   - `neuron gen`: Add `--pretty-urls` to remove `.html` suffix in generated URLs (#562)
   - Supress colors if output is not a terminal with color support (#561)
 - Performance improvements: